Update README.md
This research uses the Bank Personal Loan Modeling dataset to predict whether a customer will approve a personal loan from a bank based on customer demographic and behavioral data. The model used is a Support Vector Machine (SVM) with RBF kernel, which is optimized with SMOTE to handle class imbalance.